FIFO queue for userspace packet management
Added simple data structure and set of functions for working with this
data structure that implements a FIFO queue. This is used in conjunction
with iptables to queue packets to userspace so that we can calculate a
backlog differential between nodes. Filled in stub functions for the
packet handling thread.